refactor: rename to drizzle-test and simplify to single proxy-based approach
BREAKING CHANGE: Package renamed from pgsql-drizzle-test to drizzle-test

Major changes:
- Renamed package from pgsql-drizzle-test to drizzle-test
- Removed variation1 (wrapper-based approach) entirely
- Kept only variation2 (proxy-based approach) as the main API
- Renamed getConnectionsWithProxy to getConnections for drop-in replacement
- Re-export PgTestClient and types from pgsql-test for convenience
- Completely rewrote README to focus on proxy approach

API changes:
- import { getConnections, PgTestClient } from 'drizzle-test'
- Drop-in replacement for pgsql-test with Drizzle ORM support
- Use standard Drizzle pattern: drizzle(db.client)
- Context management via db.setContext() and db.auth()

Benefits:
- Cleaner API that matches pgsql-test conventions
- Standard Drizzle import pattern (drizzle-orm/node-postgres)
- No wrapper-specific helper methods needed
- Familiar for users of other ORMs

All 18 tests passing with the new structure
