+++ packages/drizzle-test/package.json
@@ -1,8 +1,8 @@
 {
-  "name": "pgsql-drizzle-test",
+  "name": "drizzle-test",
   "version": "0.0.1",
   "author": "Dan Lynch <pyramation@gmail.com>",
-  "description": "Drizzle ORM integration for pgsql-test - enables context-aware testing with Drizzle while maintaining RLS and transaction isolation",
+  "description": "Drop-in replacement for pgsql-test that enables Drizzle ORM to work seamlessly with context management and RLS testing",
   "main": "index.js",
   "module": "esm/index.js",
   "types": "index.d.ts",
